+++ frontend/src/app/on-time/all-operators/all-operators.component.html
@@ -6,11 +6,11 @@
       spaceAbove="2"
       spaceBelow="6"
     ></app-page-header>
-    <app-controls
+    <app-on-time-controls
       (params)="params$.next($event)"
       #controls
       [showAdminAreas]="false"
-    ></app-controls>
+    ></app-on-time-controls>
     <div class="all-operators__summary-header">
       <h2 class="govuk-heading-l">Summary</h2>
       <div class="all-operators__admin-areas">
