+++ frontend/src/app/on-time/on-time-controls/on-time-controls.component.html
@@ -1,0 +1,1 @@
+<app-controls (params)="handleParamsUpdate($event)" #controls></app-controls>
