+++ frontend/src/app/on-time/on-time-controls/on-time-controls.component.spec.ts
@@ -7,7 +7,7 @@ import {
 import { ApolloTestingModule } from "apollo-angular/testing";
 import { PerformanceFiltersInputType } from "src/generated/graphql";
 import { dateTimeEqualityMatcher } from "src/test-support/equality";
-import { ControlsComponent } from "./controls.component";
+import { ControlsComponent } from "./on-time-controls.component";
 import { FiltersComponent } from "../filters/filters.component";
 import { PanelService } from "../../shared/components/panel/panel.service";
 import { SharedModule } from "../../shared/shared.module";
