+++ frontend/src/app/on-time/view-operator/view-operator.component.html
@@ -22,10 +22,10 @@
       ></app-operator-selector>
     </div>
     <ng-container *ngIf="operator">
-      <app-controls
+      <app-on-time-controls
         [operatorId]="operator.operatorId"
         (params)="params$.next($event)"
-      ></app-controls>
+      ></app-on-time-controls>
       <app-box [padding]="['6', '5', '6', '5']" spaceAbove="6" spaceBelow="8">
         <app-tabs (tabChanged)="tabChanged($event)">
           <app-tab tabTitle="Timeline" id="timeline">
