+++ frontend/src/app/on-time/view-service/view-service.component.spec.ts
@@ -20,7 +20,7 @@ import {
 import { HeadwayService } from "../headway.service";
 import { FiltersComponent } from "../filters/filters.component";
 import { ChartNoDataWrapperComponent } from "../chart-no-data-wrapper/chart-no-data-wrapper.component";
-import { ControlsComponent } from "../controls/controls.component";
+import { ControlsComponent } from "../on-time-controls/on-time-controls.component";
 import { SharedModule } from "../../shared/shared.module";
 import { LayoutModule } from "../../layout/layout.module";
 import { OperatorService } from "../../shared/services/operator.service";
