+++ frontend/src/app/stop-analysis/view-stops/view-stops.component.html
@@ -2,9 +2,9 @@
   <ng-container ngProjectAs="content">
     <app-page-header heading="Stop Analysis"></app-page-header>
     <p>view-stops works!</p>
-    <app-controls
+    <!-- <app-controls
       (params)="params$.next($event)"
       [showAdminAreas]="false"
-    ></app-controls>
+    ></app-controls> -->
   </ng-container>
 </app-page>
